test: reorganize files
