+++ pytest/test_io_xarray.py
@@ -5,7 +5,7 @@ import xarray as xr
 
 import capytaine as cpt
 
-from capytaine.io.xarray import problems_from_dataset
+from capytaine.io.xarray import problems_from_dataset, separate_complex_values, merge_complex_values
 
 
 @pytest.fixture
@@ -54,6 +54,15 @@ def test_problems_from_dataset(sphere):
     assert cpt.RadiationProblem(body=sphere, omega=0.5, radiating_dof="Heave") in problems
     assert cpt.RadiationProblem(body=shifted_sphere, omega=0.5, radiating_dof="Heave") in problems
     assert len(problems) == 12
+
+
+def test_problems_from_dataset_incomplete_test_matrix():
+    body = cpt.FloatingBody(cpt.mesh_sphere())
+    test_matrix = xr.Dataset(coords={
+        "omega": np.linspace(0, 1, 2),
+        })
+    with pytest.raises(ValueError):
+        problems_from_dataset(test_matrix, body)
 
 
 def test_problems_from_dataset_with_wavelength(sphere):
@@ -175,6 +184,24 @@ def test_assemble_dataset(sphere, solver):
     assert "added_mass" in ds12
 
 
+def test_xarray_dataset_with_more_data():
+    # Store some mesh data when several bodies in dataset
+    bodies = [
+        cpt.FloatingBody(cpt.mesh_sphere(radius=1, resolution=(3, 3)).immersed_part(), name="sphere_1"),
+        cpt.FloatingBody(cpt.mesh_sphere(radius=2, resolution=(5, 3)).immersed_part(), name="sphere_2"),
+        cpt.FloatingBody(cpt.mesh_sphere(radius=3, resolution=(7, 3)).immersed_part(), name="sphere_3"),
+    ]
+    for body in bodies:
+        body.add_translation_dof(name="Heave")
+
+    problems = [cpt.RadiationProblem(body=b, radiating_dof="Heave", omega=1.0) for b in bodies]
+    results = cpt.BEMSolver().solve_all(problems)
+
+    ds = cpt.assemble_dataset(results, mesh=True)
+    assert 'nb_faces' in ds.coords
+    assert set(ds.coords['nb_faces'].values) == set([b.mesh.nb_faces for b in bodies])
+
+
 def test_assemble_dataset_with_infinite_free_surface(caplog, sphere, solver):
     pb = cpt.RadiationProblem(body=sphere, free_surface=np.inf)
     res = solver.solve(pb)
@@ -258,3 +285,28 @@ def test_failed_resolution_in_dataset(broken_bem_solver, sphere):
     ds = broken_bem_solver.fill_dataset(test_matrix, sphere)
     assert len(ds.wavenumber) == 5
     assert np.any(np.isnan(ds.added_mass.values))
+
+
+#######################################################################
+#                       separate_complex_values                       #
+#######################################################################
+
+def test_remove_complex_values():
+    original_dataset = xr.Dataset(
+        data_vars={
+            'variable1': (['x', 'y'], np.random.rand(4, 5)),
+            'variable2': (['x', 'y', 'z'], np.random.rand(4, 5, 3) + 1j * np.random.rand(4, 5, 3)),
+            'variable3': (['y', 'z'], np.random.rand(5, 3) + 1j * np.random.rand(5, 3)),
+        },
+        coords={
+            'x': (['x'], np.linspace(0, 10, 4))
+        })
+
+    real_dataset = separate_complex_values(original_dataset)
+    assert np.allclose(real_dataset['variable3'].sel(complex='re').data,
+                       np.real(original_dataset['variable3'].data),
+                       atol=0.1)
+    assert set(original_dataset.dims) == set(real_dataset.dims) - {'complex'}
+
+    complex_dataset = merge_complex_values(real_dataset)
+    assert set(original_dataset.dims) == set(complex_dataset.dims)
